# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Eve Online Market Browser API
  description: API for Eve Online Market Browser
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local server

paths:
  /api/esi/{endpoint}:
    get:
      summary: Proxy ESI endpoint
      description: Fetch data from ESI endpoint
      parameters:
        - name: endpoint
          in: path
          required: true
          schema:
            type: string
          description: The ESI endpoint to proxy
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token for authorization
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      tags:
        - ESI

  /api/auth/exchange-token:
    post:
      summary: Exchange authorization code for access token
      description: Exchange the authorization code for an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: Authorization code
      responses:
        '200':
          description: Successful token exchange
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      tags:
        - ESI

  /api/auth/refresh-token:
    post:
      summary: Refresh access token
      description: Refresh the access token using the refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: Refresh token
      responses:
        '200':
          description: Successful token refresh
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      tags:
        - ESI